+++ app/cmd/main.rs
@@ -8,6 +8,7 @@ async fn main() {
     Application::new()
         .config(config::register_all)
         .bootstrap(bootstrap::register)
+        .http_bootstrap(|| async { bootstrap::register_http_stack() })
         .routes(routes::register)
         .schedule(schedule::register)
         .migrations::<migrations::Migrator>()